+++ tests/test_raw_sql/util.py
@@ -5,8 +5,8 @@ from pathlib import Path
 import sqlalchemy as sa
 
 from pydiverse.pipedag import Stage, materialize
+from pydiverse.pipedag.container import RawSql
 from pydiverse.pipedag.context import ConfigContext, TaskContext
-from pydiverse.pipedag.materialize.container import RawSql
 
 
 @materialize(input_type=sa.Table, lazy=True)